+++ python/lsst/ts/wep/task/__init__.py
@@ -13,6 +13,7 @@ from .donutStamp import *
 from .donutStamps import *
 from .donutStampSelectorTask import *
 from .estimateZernikesBase import *
+from .estimateZernikesAiDonutTask import *
 from .estimateZernikesDanishTask import *
 from .estimateZernikesTieTask import *
 from .generateDonutCatalogOnlineTask import *